requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
execnet==2.1.1
Faker==28.4.1
gitig==23.4.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PySocks==1.7.1
pytest==8.3.3
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-rerunfailures==14.0
pytest-selenium==4.1.0
pytest-variables==3.1.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
ruff==0.6.5
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.0.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0